+++ src/CG.Linq.EFCore/Repositories/EFCoreRepositoryBase.cs
@@ -27,9 +27,10 @@ namespace CG.Linq.EFCore.Repositories
         #region Properties
 
         /// <summary>
-        /// This property contains the data-context associated with the repository.
+        /// This property contains a factory for creating <typeparamref name="TContext"/>
+        /// instances.
         /// </summary>
-        protected TContext DataContext { get; set; }
+        protected DbContextFactory<TContext> Factory { get; }
 
         #endregion
 
@@ -44,124 +45,19 @@ namespace CG.Linq.EFCore.Repositories
         /// class.
         /// </summary>
         /// <param name="options">The options to use with the repository.</param>
-        /// <param name="dataContext">The data-context to use with the repository.</param>
+        /// <param name="dbContextFactory">The data-context factory to use with 
+        /// the repository.</param>
         protected EFCoreRepositoryBase(
             TOptions options,
-            TContext dataContext
+            DbContextFactory<TContext> dbContextFactory
             ) 
         {
             // Validate the parameters before attempting to use them.
             Guard.Instance().ThrowIfNull(options, nameof(options))
-                .ThrowIfNull(dataContext, nameof(dataContext));
+                .ThrowIfNull(dbContextFactory, nameof(dbContextFactory));
 
             // Save the references.
-            DataContext = dataContext;
-        }
-
-        #endregion
-
-        // *******************************************************************
-        // Protected methods.
-        // *******************************************************************
-
-        #region Protected methods
-
-        /// <summary>
-        /// This method is called to clean up managed resources.
-        /// </summary>
-        /// <param name="disposing">True to cleanup managed resources.</param>
-        protected override void Dispose(
-            bool disposing
-            )
-        {
-            // Should we cleanup managed resources?
-            if (disposing)
-            {
-                DataContext?.Dispose();
-            }
-
-            // Give the base class a chance.
-            base.Dispose(disposing);
-        }
-
-        #endregion
-    }
-
-
-
-    /// <summary>
-    /// This class is a base EFCORE implementation of the <see cref="ILinqRepository{TModel}"/>
-    /// interface.
-    /// </summary>
-    /// <typeparam name="TContext">The data-context type associated with the repository.</typeparam>
-    /// <typeparam name="TOptions">The options type associated with the repository.</typeparam>
-    /// <typeparam name="TModel">The model type associated with the repository.</typeparam>
-    public abstract class EFCoreRepositoryBase<TContext, TOptions, TModel> :
-        EFCoreRepositoryBase<TContext, TOptions>,
-        ILinqRepository<TModel>
-        where TModel : class, IModel
-        where TOptions : IOptions<EFCoreRepositoryOptions>
-        where TContext : DbContext
-    {
-        // *******************************************************************
-        // Constructors.
-        // *******************************************************************
-
-        #region Constructors
-
-        /// <summary>
-        /// This constructor creates a new instance of the <see cref="EFCoreRepositoryBase{TContext, TOptions, TModel}"/>
-        /// class.
-        /// </summary>
-        /// <param name="options">The options to use with the repository.</param>
-        /// <param name="dataContext">The data-context to use with the repository.</param>
-        protected EFCoreRepositoryBase(
-            TOptions options,
-            TContext dataContext
-            ) : base(options, dataContext)
-        {
-
-        }
-
-        #endregion
-
-        // *******************************************************************
-        // Public methods.
-        // *******************************************************************
-
-        #region Public methods
-
-        /// <inheritdoc />
-        public virtual IQueryable<TModel> AsQueryable()
-        {
-            // Defer to the data-context.
-            return DataContext.Set<TModel>().AsQueryable();
-        }
-
-        #endregion
-
-        // *******************************************************************
-        // Protected methods.
-        // *******************************************************************
-
-        #region Protected methods
-
-        /// <summary>
-        /// This method is called to clean up managed resources.
-        /// </summary>
-        /// <param name="disposing">True to cleanup managed resources.</param>
-        protected override void Dispose(
-            bool disposing
-            )
-        {
-            // Should we cleanup managed resources?
-            if (disposing)
-            {
-                DataContext?.Dispose();
-            }
-
-            // Give the base class a chance.
-            base.Dispose(disposing);
+            Factory = dbContextFactory;
         }
 
         #endregion